Moves pages to their own folder
