+++ voting/client-web/src/pages/DelegatorsPage/GuardiansList.tsx
@@ -17,8 +17,8 @@ import TableHead from '@material-ui/core/TableHead';
 import TableRow from '@material-ui/core/TableRow';
 import React, { useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { CopyAddressButton } from '../CopyAddressButton/CopyAddressButton';
-import { VoteChip } from '../VoteChip/VoteChip';
+import { CopyAddressButton } from '../../components/CopyAddressButton/CopyAddressButton';
+import { VoteChip } from '../../components/VoteChip/VoteChip';
 import { TGuardianInfoExtended } from '../../Store/GuardiansStore';
 import { observable } from 'mobx';
 import { observer } from 'mobx-react';
