+++ voting/client-web/src/pages/DelegatorsPage/delegateButton.tsx
