+++ voting/client-web/src/pages/RewardsPage/DelegationInfoTable.tsx
@@ -7,7 +7,7 @@
  */
 
 import React, { useMemo } from 'react';
-import { VoteChip } from '../VoteChip/VoteChip';
+import { VoteChip } from '../../components/VoteChip/VoteChip';
 import Table from '@material-ui/core/Table';
 import TableRow from '@material-ui/core/TableRow';
 import TableBody from '@material-ui/core/TableBody';
