+++ voting/client-web/src/pages/RewardsPage/RewardsPage.tsx
