+++ voting/client-web/src/pages/ValidatorsPage/ValidatorsPage.tsx
